requirements.txt
pyserial>=3.5 
tabulate>=0.9.0 